Fixed small bug related to the password generator
if no property is checked and if we click on the generate password it
r slide the slider of password length call the generate password button
and in the password field undefined is printed repeatively (password length times).

so I used the useRef for referencing the button and whenever we removes all the checks than button and slider is disabled.

I updated the ui as well